Bugfix: The Option "onlyDirectoryDependencies"
When the Options `onlyDirectDependencies` was provided, the script always took the `dependencies`, even if the flag `devDependencies` was provided. This Change gives you only the direct development dependencies if the flags `onlyDirectDependencies` & `devDependencies` ist provided. When `onlyDirectDependencies` is the provided without the `devDependencies` Options, it will return the normal dependencies.